U1cMPO

U1cMPO.cmpo

U1cMPO.cmps

U1cMPO.u1_matrix

Base.:*

Base.:*

Base.:*

Base.:*

Base.:*

Base.:*

Base.:*

Base.:+

Base.:-

Base.:==

Base.adjoint

Base.isapprox

Base.kron

Base.transpose

Base.vec

ChainRulesCore.rrule

ChainRulesCore.rrule

LinearAlgebra.tr

U1cMPO.:⊗

U1cMPO.C_matrix

U1cMPO.L2_theta_0

U1cMPO.L2_theta_pi

U1cMPO.X_theta_0

U1cMPO.X_theta_pi

U1cMPO.add_to_submat!

U1cMPO.add_to_submat!

U1cMPO.blk_chk

U1cMPO.d_vec_0

U1cMPO.d_vec_pi

U1cMPO.diagQ

U1cMPO.eigh

U1cMPO.energy

U1cMPO.eye_u1m

U1cMPO.eye_u1m

U1cMPO.eye_u1m

U1cMPO.f_and_gf_gen

U1cMPO.fidel_and_gfidel_gen

U1cMPO.free_energy

U1cMPO.init_diag_u1m

U1cMPO.init_u1m_by_params

U1cMPO.isometry_truncate

U1cMPO.klein

U1cMPO.kron_bookkeeping

U1cMPO.l_labels_f

U1cMPO.log_tr_expm

U1cMPO.logsumexp

U1cMPO.ovlp

U1cMPO.params_len

U1cMPO.pauli

U1cMPO.pauli_id

U1cMPO.reduced_density_matrix

U1cMPO.reflect

U1cMPO.reflect

U1cMPO.submat

U1cMPO.symmetrize

U1cMPO.toarray

U1cMPO.trunc_msk_gen

U1cMPO.truncate_according_to_rdm

U1cMPO.von_neumann_entropy

U1cMPO.zero_u1m

U1cMPO.cmpo — Type

cMPO (for now, only nearest-neighboring (NN) interactions)

\[\left( \begin{array}{cc} Q & \mathbf{R}\\ \mathbf{L} & 0 \end{array} \right)\]

U1cMPO.cmps — Type

cMPS

U1cMPO.u1_matrix — Type

definition for matrices with U(1) block structures.

Base.:* — Method

*(x, m)

Number-Matrix multiplication.

Base.:* — Method

*(op1, op2)

Multiply cMPO op1 with cMPO op2.

Base.:* — Method

*(op, psi)

Act the cMPO op onto the cMPS psi, and return the new cMPS.

Base.:* — Method

*(phi, op)

Act the cMPO op to the left cMPS phi.

Base.:* — Method

*(phi, psi)

Return the K-matrix from the input cMPS phi and psi.

Base.:* — Method

*(m, x)

Number-Matrix multiplication.

Base.:* — Method

*(m1, m2)

Matrix multiplication.

Base.:+ — Method

+(m1, m2)

Add two matrices.

Base.:- — Method

-(m1, m2)

Calculate m1 - m2.

Base.:== — Method

==(m1, m2)

Return whether two matrices are identical.

Base.adjoint — Method

adjoint(m)

This function does nothing to the matrix. It is useful when calculating the adjoint (transpose) of a cMPO.

Base.isapprox — Method

isapprox(m1, m2)

Return whether two matrices are approximately identical.

Base.kron — Method

kron(m1, m2)

Tensor product of matrices.

Base.transpose — Method

transpose(m)

Take the transpose of a u1_matrix.

Base.vec — Method

vec(u1_matrix, is_diag=false)

Convert all the parameters (the data within the blocks) to a vector. If is_diag is set to true, then only take the diagonal elements.

ChainRulesCore.rrule — Method

rules for U1Matrix

ChainRulesCore.rrule — Method

rules missed in LinearAlgebra

LinearAlgebra.tr — Method

tr(m)

Trace of matrix.

U1cMPO.:⊗ — Method

⊗(m1, m2)

Tensor product of matrices.

U1cMPO.C_matrix — Method

C_matrix(psi, beta, tau)

A util function for the the calculation of the von-Neumann entropy which calculates the coefficent matrix. psi is cMPS, beta is the inversed temperature, and tau is the interval in the imaginary time. We have inserted some unitary matrices to maintain the U(1) block structure of the matrix.

U1cMPO.L2_theta_0 — Method

L2_theta_0(lmax)

Return the matrix representation for the kinetic term in quantum rotor model (no theta term).

U1cMPO.L2_theta_pi — Method

L2_theta_0(lmax)

Return the matrix representation for the kinetic term in quantum rotor model (theta=pi).

U1cMPO.X_theta_0 — Method

X_theta_0(M, lmax)

Return matrix representation of spherical tensor operator for ordinary quantum rotors.

U1cMPO.X_theta_pi — Method

X_theta_pi(M, lmax)

Return matrix representation of spherical tensor operator for quantum rotors decorated with a magnetic monopole (theta=pi).

U1cMPO.add_to_submat! — Method

add_to_submat!(m, subm_to_add, l_label, l_indices, r_indices)

Inplace modification of the U1 matrix m by adding some small matrix subm_to_add to it. The small matrix is added to the block corresponding to the quantum number l_label, within the sub-block indicated by the index ranges l_indices and r_indices.

U1cMPO.add_to_submat! — Method

add_to_submat!(m, subm_to_add, l_label)

Inplace modification of the U1 matrix m by adding some small matrix subm_to_add to it. The small matrix is added to the whole block corresponding to the quantum number l_label.

U1cMPO.blk_chk — Method

blk_chk(m1, m2)

Return whether two matrices has the same block structure.

U1cMPO.d_vec_0 — Method

d_vec_0(lmax)

Return the U(1) block structure given the largest l kept in the case of no theta term.

U1cMPO.d_vec_pi — Method

d_vec_pi(lmax)

Return the U(1) block structure given the largest l kept in the case of theta=pi.

U1cMPO.diagQ — Method

diagQ(psi)

Perform gauge transformation to cMPS psi so that its Q matrix is diagonalized.

U1cMPO.eigh — Method

eigh(m)

Eigendecomposition of a (symmetric) U1 matrix.

U1cMPO.energy — Method

energy(T, psi, beta)

Calculate energy from cMPO T, cMPS psi at inversed temperature beta.

U1cMPO.eye_u1m — Method

eye_u1m(T, d_vec)

Initialize an identity matrix of datatype T, and assign a block structure to it.

U1cMPO.eye_u1m — Method

eye_u1m(op)

Given a cMPO, generate an identity matrix according to its U(1) quantum number sectors.

U1cMPO.eye_u1m — Method

eye_u1m(psi)

Given a cMPS, generate an identity matrix according to its U(1) quantum number sectors.

U1cMPO.f_and_gf_gen — Method

f_and_gf_gen(T, params_to_cmps, beta)

Given a function params_to_cmps that convert parameter vector to cmps, the Hermtian cMPO T, and the inversed temperature beta, generate functions that calculate free energy and the corresponding gradient from parameters.

U1cMPO.fidel_and_gfidel_gen — Method

fidel_and_gfidel_gen(phi, params_to_cmps, beta)

Given a function params_to_cmps that convert parameter vector to cmps, the target cMPS phi, and the inversed temperature beta, generate functions that calculate free energy and the corresponding gradient from parameters.

U1cMPO.free_energy — Method

free_energy(T, psi, beta)

given cMPS psi, Hermitian cMPO T and the inversed temperature beta, calculate the free energy.

U1cMPO.init_diag_u1m — Method

init_diag_u1m(d_vec, params)

Initialize a diagonal matrix from a vector params, and then assign a block structure to it.

U1cMPO.init_u1m_by_params — Method

init_u1m_by_params(d_vec, qn, params)

An alternative initialization for u1_matrix from a vector of parameters.

U1cMPO.isometry_truncate — Method

isometry_truncate(m, v, msk_vec)

Truncate the U1 matrix m by an isometry matrix. We implement this by first transform the matrix by a orthogonal matrix v that has the same block structure with m, and then truncate the transformed U1 matrix according to the mask msk_vec, block by block.

U1cMPO.klein — Method

klein(Lpsi, psi, beta)

Calculate the klein bottle entropy from left cMPS Lpsi and right cMPS psi at inversed temperature beta.

U1cMPO.kron_bookkeeping — Method

kron_bookkeeping(d_vec1, d_vec2)

Utilization function for the calculation of the tensor product. Given two U(1) quantum number sectors, return

the new quantum number sector for the tensor product matrix.

a function that maps the qunatum numbers (n1, n2) corresponding to input quantum number sectors to the sub-block location (l, x0, x1) within the new matrix, where l is the new quantum number and x0 and x1 represent the beginning and ending point within in the block.

U1cMPO.l_labels_f — Method

l_labels_f(nlen, qn)

Given the total number nlen of the U(1) quantum numbers and the quantum number qn of the U1 matrix, return a vector that maps the index of the block to the U(1) quantum number of the block. For example, when qn=1, the first block corresponds to quantum number 1. when qn=-1, however, the first block corresponds to quantum number 2. (The U(1) quantum number is assumed to start from 1 to nlen.)

U1cMPO.log_tr_expm — Method

log_tr_expm(m, beta)

Calculates

\[\log \mathrm{Tr} \exp(\beta M)\]

U1cMPO.logsumexp — Method

logsumexp(w)

Calculation of logsumexp.

U1cMPO.ovlp — Method

ovlp(phi, psi, beta)

Given inversed temperature beta, calculate the log of the overlap between the cMPSs phi and psi

U1cMPO.params_len — Method

params_len(d_vec, qn)

Given the quantum number sector d_vec and the quantum number of the matrix qn, determine the total number of the parameters.

U1cMPO.pauli — Method

pauli(M)

Generate Pauli matrix (σ0, σ+, and σ-).

U1cMPO.pauli_id — Method

pauli_id()

Generate 2×2 identity matrix.

U1cMPO.reduced_density_matrix — Method

reduced_density_matrix(psi, beta)

For cMPS psi with imaginary-time length beta, construct the reduced density matrix. We have inserted some unitaries to maintain the U(1) block structure of the matrix.

U1cMPO.reflect — Method

reflect(psi)

reflect the matrices contained in the input cMPS psi. See also: reflect(m)

U1cMPO.reflect — Method

reflect(m)

reflect a u1_matrix. Equivalent to the following unitary tranformation

\[\left( \begin{array}{cccc} & & & 1\\ & & 1 & \\ & \udots & & \\ 1 & & & \end{array} \right)\]

U1cMPO.submat — Method

submat(m, n)

Return the block corresponding to U(1) quantum number n.

U1cMPO.symmetrize — Method

symmetrize(m)

Symmetrize the matrix by

\[(M + M^T) / 2\]

U1cMPO.toarray — Method

to_array(m)

Represent the U1 matrix as an ordinary 2D-array.

U1cMPO.trunc_msk_gen — Method

trunc_msk_gen(d_vec, w, num_kept)

Generate msk_vec for the function isometry_truncate(m, v, msk_vec). According to the block structure d_vec and some spectrum w, keep num_kept of states in the spectrum. Return a vector of BitVector that tells how to truncate a U1 matrix, block by block.

U1cMPO.truncate_according_to_rdm — Method

truncate_according_to_rdm(psi, chi, beta)

Truncate cMPS psi to target bond dimension chi at inversed temperature beta according to the reduced density matrix eigenvalues. The new U(1) quantum number sectors can be automatically determined from this process.

U1cMPO.von_neumann_entropy — Method

von_neumann_entropy(psi, beta, tau)

For cMPS psi, calculate the von-Neumann entropy between intervals τ and β-τ.

U1cMPO.zero_u1m — Method

zero_u1m(T, d_vec)

Initialize a zero matrix of datatype T, and assign a block structure to it.